VJOURNAL: Split code into two files.
Code is getting too dense to be in one file only.
